backend/requirements.txt
# Core dependencies for Driver Drowsiness Detection System

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3

# Data Processing
pandas==2.1.4
pillow==10.1.0

# Testing
pytest==7.4.3
hypothesis==6.92.1

# Utilities
scipy==1.11.4
matplotlib==3.8.2
psutil==5.9.6
tqdm==4.66.1

# Security and Privacy
cryptography==41.0.7
